Kernels

GaussianProcesses.Const — Type

Const <: Kernel

Constant kernel

\[k(x,x') = σ²\]

with signal standard deviation $σ$.

GaussianProcesses.Const — Method

Constant kernel function

Const(lσ::T)

Arguments

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Lin — Method

Lin(ll::Union{Real,Vector{Real}})

Create linear kernel with length scale exp.(ll).

See also LinIso and LinArd.

GaussianProcesses.LinArd — Type

LinArd <: Kernel

ARD linear kernel (covariance)

\[k(x,x') = xᵀL⁻²x'\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.LinArd — Method

Linear ARD Covariance Function

LinArd(ll::Vector{T})

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

GaussianProcesses.LinIso — Type

LinIso <: Kernel

Isotropic linear kernel (covariance)

\[k(x, x') = xᵀx'/ℓ²\]

with length scale $ℓ$.

GaussianProcesses.LinIso — Method

Linear Isotropic Covariance Function

LinIso(ll::T)

Arguments

ll::Real: length scale (given on log scale)

GaussianProcesses.Masked — Type

Masked{K<:Kernel} <: Kernel

A wrapper for kernels so that they are only applied along certain dimensions.

This is similar to the active_dims kernel attribute in the python GPy package and to the covMask function in the matlab gpml package.

The implementation is very simple: any function of the kernel that takes an X::Matrix input is delegated to the wrapped kernel along with a view of X that only includes the active dimensions.

GaussianProcesses.Matern — Method

Matern(ν::Real, ll::Union{Real,Vector{Real}}, lσ::Real)

Create Matérn kernel of type ν (i.e. ν = 1/2, ν = 3/2, or ν = 5/2) with length scale exp.(ll) and signal standard deviation exp(σ).

See also Mat12Iso, Mat12Ard, Mat32Iso, Mat32Ard, Mat52Iso, and Mat52Ard.

GaussianProcesses.Mat12Ard — Type

Mat12Ard <: MaternARD

ARD Matern 1/2 kernel (covariance)

\[k(x,x') = σ² \exp(-|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.Mat12Ard — Method

Matern 1/2 ARD covariance Function

Mat12Ard(ll::Vector{T}, lσ::T)

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Mat12Iso — Type

Mat12Iso <: MaternISO

Isotropic Matern 1/2 kernel (covariance)

\[k(x,x') = σ^2 \exp(-|x-y|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

GaussianProcesses.Mat12Iso — Method

Matern 1/2 isotropic covariance Function

Mat12Iso(ll::T, lσ::T)

Arguments

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Mat32Ard — Type

Mat32Ard <: MaternARD

ARD Matern 3/2 kernel (covariance)

\[k(x,x') = σ²(1 + √3|x-x'|/L)\exp(- √3|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.Mat32Ard — Method

Matern 3/2 ARD covariance function

Mat32Ard(ll::Vector{T}, lσ::T)

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Mat32Iso — Type

Mat32Iso <: MaternIso

Isotropic Matern 3/2 kernel (covariance)

\[k(x,x') = σ²(1 + √3|x-x'|/ℓ)\exp(-√3|x-x'|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

GaussianProcesses.Mat32Iso — Method

Matern 3/2 isotropic covariance function

Mat32Iso(ll::T, lσ::T)

Arguments

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Mat52Ard — Type

Mat52Ard <: MaternARD

ARD Matern 5/2 kernel (covariance)

\[k(x,x') = σ²(1 + √5|x-x'|/L + 5|x-x'|²/(3L²))\exp(- √5|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.Mat52Ard — Method

Matern 5/2 ARD covariance Function

Mat52Ard(ll::Vector{Real}, lσ::Real)

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Mat52Iso — Type

Mat52Iso <: MaternIso

Isotropic Matern 5/2 kernel (covariance)

\[k(x,x') = σ²(1+√5|x-x'|/ℓ + 5|x-x'|²/(3ℓ²))\exp(- √5|x-x'|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

GaussianProcesses.Mat52Iso — Method

Matern 5/2 isotropic covariance function

Mat52Iso(ll::Real, lσ::Real)

Arguments

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Noise — Type

Noise <: Kernel

Noise kernel (covariance)

\[k(x,x') = σ²δ(x-x'),\]

where $δ$ is the Kronecker delta function and $σ$ is the signal standard deviation.

GaussianProcesses.Noise — Method

White Noise kernel

Noise(lσ::Real)

Arguments

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.Periodic — Type

Periodic <: Isotropic{Euclidean}

Periodic kernel (covariance)

\[k(x,x') = σ²\exp(-2\sin²(π|x-x'|/p)/ℓ²)\]

with length scale $ℓ$, signal standard deviation $σ$, and period $p$.

GaussianProcesses.Periodic — Method

Periodic kernel function

Periodic(ll::Real, lσ::Real, lp::Real)

Arguments

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

lp::Real: periodicity parameter (given on log scale)

GaussianProcesses.Poly — Type

Poly <: Kernel

Polynomial kernel (covariance)

\[k(x,x') = σ²(xᵀx' + c)ᵈ\]

with signal standard deviation $σ$, additive constant $c$, and degree $d$.

GaussianProcesses.Poly — Method

Polynomial kernel function

Poly(lc::Real, lσ::Real, deg::Int)

Arguments

lc::Real: additive constant (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

deg::Int: degree of polynomial

GaussianProcesses.RQ — Method

RQ(ll::Union{Real,Vector{Real}}, lσ::Real, lα::Real)

Create Rational Quadratic kernel with length scale exp.(ll), signal standard deviation exp(lσ), and shape parameter exp(lα).

See also RQIso and RQArd.

GaussianProcesses.RQArd — Type

RQArd <: StationaryARD{WeightedSqEuclidean}

ARD Rational Quadratic kernel (covariance)

\[k(x,x') = σ²(1 + (x - x')ᵀL⁻²(x - x')/(2α))^{-α}\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$, signal standard deviation $σ$, and shape parameter $α$ where $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.RQArd — Method

Rational Quadratic ARD Covariance Function

RQArd(ll::Vector{Real}, lσ::Real, lα::Real)

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

lα::Real: shape parameter (given on log scale)

GaussianProcesses.RQIso — Type

RQIso <: Isotropic{SqEuclidean}

Isotropic Rational Quadratic kernel (covariance)

\[k(x,x') = σ²(1 + (x - x')ᵀ(x - x')/(2αℓ²))^{-α}\]

with length scale $ℓ$, signal standard deviation $σ$, and shape parameter $α$.

GaussianProcesses.RQIso — Method

Rational Quadratic Isotropic Covariance Function

RQIso(ll:T, lσ::T, lα::T)

Arguments

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

lα::Real: shape parameter (given on log scale)

GaussianProcesses.SE — Method

SE(ll::Union{Real,Vector{Real}}, lσ::Real)

Create squared exponential kernel with length scale exp.(ll) and signal standard deviation exp(lσ).

See also SEIso and SEArd.

GaussianProcesses.SEArd — Type

SEArd <: StationaryARD{WeightedSqEuclidean}

ARD Squared Exponential kernel (covariance)

\[k(x,x') = σ²\exp(- (x - x')ᵀL⁻²(x - x')/2)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

GaussianProcesses.SEArd — Method

Squared Exponential Function with ARD

SEArd(ll::Vector{Real}, lσ::Real)

Arguments

ll::Vector{Real}: vector of length scales (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.SEIso — Type

SEIso <: Isotropic{SqEuclidean}

Isotropic Squared Exponential kernel (covariance)

\[k(x,x') = σ²\exp(- (x - x')ᵀ(x - x')/(2ℓ²))\]

with length scale $ℓ$ and signal standard deviation $σ$.

GaussianProcesses.SEIso — Method

Squared Exponential kernel function

SEIso(ll::T, lσ::T)

Arguments:

ll::Real: length scale (given on log scale)

lσ::Real: signal standard deviation (given on log scale)

GaussianProcesses.EmptyData — Type

EmptyData <: KernelData

Default empty KernelData.

GaussianProcesses.KernelData — Type

KernelData

Data to be used with a kernel object to calculate a covariance matrix, which is independent of kernel hyperparameters.

See also EmptyData.

GaussianProcesses.cov! — Function

cov!(cK::AbstractMatrix, k::Kernel, X1::AbstractMatrix, X2::AbstractMatrix)

Like cov(k, X1, X2), but stores the result in cK rather than a new matrix.

Statistics.cov — Function

cov(k::Kernel, X::AbstractMatrix[, data::KernelData = EmptyData()])

Create covariance matrix from kernel k, matrix of observations X, where each column is an observation, and kernel data data constructed from input observations.

Statistics.cov — Function

cov(k::Kernel, X1::AbstractMatrix, X2::AbstractMatrix)

Create covariance matrix from kernel k and matrices of observations X1 and X2, where each column is an observation.