Gaussian Processes

Gaussian Processes

GaussianProcesses.predict_f — Method.

predict_f(gp::GPBase, X::Matrix{Float64}[; full_cov::Bool = false])

Return posterior mean and variance of the Gaussian Process gp at specfic points which are given as columns of matrix X. If full_cov is true, the full covariance matrix is returned instead of only variances.

source

GaussianProcesses.GPE — Method.

GPE(x, y, mean, kernel[, logNoise])

Fit a Gaussian process to a set of training points. The Gaussian process is defined in terms of its user-defined mean and covariance (kernel) functions. As a default it is assumed that the observations are noise free.

Arguments:

x::AbstractVecOrMat{Float64}: Input observations

y::AbstractVector{Float64}: Output observations

mean::Mean: Mean function

kernel::Kernel: Covariance function

logNoise::Float64: Natural logarithm of the standard deviation for the observation noise. The default is -2.0, which is equivalent to assuming no observation noise.

source

GaussianProcesses.GPE — Method.

GPE(; mean::Mean = MeanZero(), kernel::Kernel = SE(0.0, 0.0), logNoise::Float64 = -2.0)

Construct a GPE object without observations.

source

GaussianProcesses.GP — Function.

GP(x, y, mean::Mean, kernel::Kernel[, logNoise::Float64=-2.0])

Fit a Gaussian process that is defined by its mean, its kernel, and the logarithm logNoise of the standard deviation of its observation noise to a set of training points x and y.

See also: GPE

source

GaussianProcesses.predict_y — Method.

predict_y(gp::GPE, x::Union{Vector{Float64},Matrix{Float64}}[; full_cov::Bool=false])

Return the predictive mean and variance of Gaussian Process gp at specfic points which are given as columns of matrix x. If full_cov is true, the full covariance matrix is returned instead of only variances.

source

GaussianProcesses.update_target! — Method.

update_target!(gp::GPE, ...)

Update the log-posterior

\[\log p(θ | y) ∝ \log p(y | θ) + \log p(θ)\]

of a Gaussian process gp.

source

GaussianProcesses.GPMC — Method.

GPMC(x, y, mean, kernel, lik)

Fit a Gaussian process to a set of training points. The Gaussian process with non-Gaussian observations is defined in terms of its user-defined likelihood function, mean and covaiance (kernel) functions.

The non-Gaussian likelihood is handled by a Monte Carlo method. The latent function values are represented by centered (whitened) variables $f(x) = m(x) + Lv$ where $v ∼ N(0, I)$ and $LLᵀ = K_θ$.

Arguments:

x::AbstractVecOrMat{Float64}: Input observations

y::AbstractVector{<:Real}: Output observations

mean::Mean: Mean function

kernel::Kernel: Covariance function

lik::Likelihood: Likelihood function

source

GaussianProcesses.GP — Method.

GP(x, y, mean::Mean, kernel::Kernel, lik::Likelihood)

Fit a Gaussian process that is defined by its mean, its kernel, and its likelihood function lik to a set of training points x and y.

See also: GPMC

source

GaussianProcesses.predict_y — Method.

predict_y(gp::GPMC, x::Union{Vector{Float64},Matrix{Float64}}[; full_cov::Bool=false])

Return the predictive mean and variance of Gaussian Process gp at specfic points which are given as columns of matrix x. If full_cov is true, the full covariance matrix is returned instead of only variances.

source

GaussianProcesses.update_target! — Method.

update_target!(gp::GPMC, ...)

Update the log-posterior

\[\log p(θ, v | y) ∝ \log p(y | v, θ) + \log p(v) + \log p(θ)\]

of a Gaussian process gp.

source

GaussianProcesses.mcmc — Method.

mcmc(gp::GPBase; kwargs...)

Run MCMC algorithms provided by the Klara package for estimating the hyperparameters of Gaussian process gp.

source

GaussianProcesses.make_posdef! — Method.

make_posdef!(m::Matrix{Float64}, chol_factors::Matrix{Float64})

Try to encode covariance matrix m as a positive definite matrix. The chol_factors matrix is recycled to store the cholesky decomposition, so as to reduce the number of memory allocations.

Sometimes covariance matrices of Gaussian processes are positive definite mathematically but have negative eigenvalues numerically. To resolve this issue, small weights are added to the diagonal (and hereby all eigenvalues are raised by that amount mathematically) until all eigenvalues are positive numerically.

source

GaussianProcesses.fit! — Method.

fit!(gp::GPE{X,Y}, x::X, y::Y)

Fit Gaussian process GPE to a training data set consisting of input observations x and output observations y.

source

GaussianProcesses.get_ααinvcKI! — Method.

get_ααinvcKI!(ααinvcKI::Matrix{Float64}, cK::AbstractPDMat, α::Vector)

Write ααᵀ - cK⁻¹ to ααinvcKI avoiding any memory allocation, where cK and α are the covariance matrix and the alpha vector of a Gaussian process, respectively. Hereby α is defined as cK⁻¹ (Y - μ).

source

GaussianProcesses.initialise_target! — Method.

initialise_target!(gp::GPE)

Initialise the log-posterior

\[\log p(θ | y) ∝ \log p(y | θ) + \log p(θ)\]

of a Gaussian process gp.

source

GaussianProcesses.update_cK! — Method.

update_cK!(gp::GPE)

Update the covariance matrix and its Cholesky decomposition of Gaussian process gp.

source

GaussianProcesses.update_dmll! — Method.

 update_dmll!(gp::GPE, ...)

Update the gradient of the marginal log-likelihood of Gaussian process gp.

source

GaussianProcesses.update_mll_and_dmll! — Method.

update_mll_and_dmll!(gp::GPE, ...)

Update the gradient of the marginal log-likelihood of a Gaussian process gp.

source

GaussianProcesses.update_target_and_dtarget! — Method.

update_target_and_dtarget!(gp::GPE, ...)

Update the log-posterior

\[\log p(θ | y) ∝ \log p(y | θ) + \log p(θ)\]

of a Gaussian process gp and its derivative.

source

GaussianProcesses.initialise_ll! — Method.

initialise_ll!(gp::GPMC)

Initialise the log-likelihood of Gaussian process gp.

source

GaussianProcesses.initialise_target! — Method.

initialise_target!(gp::GPMC)

Initialise the log-posterior

\[\log p(θ, v | y) ∝ \log p(y | v, θ) + \log p(v) + \log p(θ)\]

of a Gaussian process gp.

source

GaussianProcesses.update_cK! — Method.

update_cK!(gp::GPMC)

Update the covariance matrix and its Cholesky decomposition of Gaussian process gp.

source

GaussianProcesses.update_dll! — Method.

 update_dll!(gp::GPMC, ...)

Update the gradient of the log-likelihood of Gaussian process gp.

source

GaussianProcesses.update_target_and_dtarget! — Method.

update_target_and_dtarget!(gp::GPMC, ...)

Update the log-posterior

\[\log p(θ, v | y) ∝ \log p(y | v, θ) + \log p(v) + \log p(θ)\]

of a Gaussian process gp and its derivative.

source

GaussianProcesses.composite_param_names — Method.

composite_param_names(objects, prefix)

Call get_param_names on each element of objects and prefix the returned name of the element at index i with prefix * i * '_'.

Examples

julia> GaussianProcesses.get_param_names(ProdKernel(Mat12Iso(1/2, 1/2), SEArd([0.0, 1.0], 0.0)))
5-element Array{Symbol,1}:
 :pk1_ll
 :pk1_lσ
 :pk2_ll_1
 :pk2_ll_2
 :pk2_lσ

source

GaussianProcesses.map_column_pairs! — Method.

map_column_pairs!(D::Matrix{Float64}, f, X::Matrix{Float64}[, Y::Matrix{Float64} = X])

Like map_column_pairs, but stores the result in D rather than a new matrix.

source

GaussianProcesses.map_column_pairs — Method.

map_column_pairs(f, X::Matrix{Float64}[, Y::Matrix{Float64} = X])

Create a matrix by applying function f to each pair of columns of input matrices X and Y.

source