Kernels

Kernels

GaussianProcesses.Const — Type.

Const <: Kernel

Constant kernel

\[k(x,x') = σ²\]

with signal standard deviation $σ$.

source

GaussianProcesses.Const — Method.

Const(lσ::T)

Create Const with signal standard deviation exp(lσ).

source

GaussianProcesses.Lin — Method.

Lin(ll::Union{Real,Vector{Real}})

Create linear kernel with length scale exp.(ll).

See also LinIso and LinArd.

source

GaussianProcesses.LinArd — Type.

LinArd <: Kernel

ARD linear kernel (covariance)

\[k(x,x') = xᵀL⁻²x'\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.LinArd — Method.

LinArd(ll::Vector{T})

Create LinArd with length scale exp.(ll).

source

GaussianProcesses.LinIso — Type.

LinIso <: Kernel

Isotropic linear kernel (covariance)

\[k(x, x') = xᵀx'/ℓ²\]

with length scale $ℓ$.

source

GaussianProcesses.LinIso — Method.

LinIso(ll::T)

Create LinIso with length scale exp(ll).

source

GaussianProcesses.Masked — Type.

Masked{K<:Kernel} <: Kernel

A wrapper for kernels so that they are only applied along certain dimensions.

This is similar to the active_dims kernel attribute in the python GPy package and to the covMask function in the matlab gpml package.

The implementation is very simple: any function of the kernel that takes an X::Matrix input is delegated to the wrapped kernel along with a view of X that only includes the active dimensions.

source

GaussianProcesses.Matern — Method.

Matern(ν::Real, ll::Union{Real,Vector{Real}}, lσ::Real)

Create Matérn kernel of type ν (i.e. ν = 1/2, ν = 3/2, or ν = 5/2) with length scale exp.(ll) and signal standard deviation exp(σ).

See also Mat12Iso, Mat12Ard, Mat32Iso, Mat32Ard, Mat52Iso, and Mat52Ard.

source

GaussianProcesses.Mat12Ard — Type.

Mat12Ard <: MaternARD

ARD Matern 1/2 kernel (covariance)

\[k(x,x') = σ² \exp(-|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.Mat12Ard — Method.

Mat12Ard(ll::Vector{T}, lσ::T)

Create Mat12Ard with length scale exp.(ll) and signal standard deviation exp(σ).

source

GaussianProcesses.Mat12Iso — Type.

Mat12Iso <: MaternISO

Isotropic Matern 1/2 kernel (covariance)

\[k(x,x') = σ^2 \exp(-|x-y|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

source

GaussianProcesses.Mat12Iso — Method.

Mat12Iso(ll::T, lσ::T)

Create Mat12Iso with length scale exp(ll) and signal standard deviation exp(σ).

source

GaussianProcesses.Mat32Ard — Type.

Mat32Ard <: MaternARD

ARD Matern 3/2 kernel (covariance)

\[k(x,x') = σ²(1 + √3|x-x'|/L)\exp(- √3|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.Mat32Ard — Method.

Mat32Ard(ll::Vector{T}, lσ::T)

Create Mat32Ard with length scale exp.(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.Mat32Iso — Type.

Mat32Iso <: MaternIso

Isotropic Matern 3/2 kernel (covariance)

\[k(x,x') = σ²(1 + √3|x-x'|/ℓ)\exp(-√3|x-x'|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

source

GaussianProcesses.Mat32Iso — Method.

Mat32Iso(ll::T, lσ::T)

Create Mat32Iso with length scale exp(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.Mat52Ard — Type.

Mat52Ard <: MaternARD

ARD Matern 5/2 kernel (covariance)

\[k(x,x') = σ²(1 + √5|x-x'|/L + 5|x-x'|²/(3L²))\exp(- √5|x-x'|/L)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.Mat52Ard — Method.

Mat52Ard(ll::Vector{Real}, lσ::Real)

Create Mat52Ard with length scale exp.(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.Mat52Iso — Type.

Mat52Iso <: MaternIso

Isotropic Matern 5/2 kernel (covariance)

\[k(x,x') = σ²(1+√5|x-x'|/ℓ + 5|x-x'|²/(3ℓ²))\exp(- √5|x-x'|/ℓ)\]

with length scale $ℓ$ and signal standard deviation $σ$.

source

GaussianProcesses.Mat52Iso — Method.

Mat52Iso(ll::Real, lσ::Real)

Create Mat52Iso with length scale exp(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.Noise — Type.

Noise <: Kernel

Noise kernel (covariance)

\[k(x,x') = σ²δ(x-x'),\]

where $δ$ is the Kronecker delta function and $σ$ is the signal standard deviation.

source

GaussianProcesses.Noise — Method.

Noise(lσ::Real)

Create Noise with signal standard deviation exp(lσ).

source

GaussianProcesses.Periodic — Type.

Periodic <: Isotropic{Euclidean}

Periodic kernel (covariance)

\[k(x,x') = σ²\exp(-2\sin²(π|x-x'|/p)/ℓ²)\]

with length scale $ℓ$, signal standard deviation $σ$, and period $p$.

source

GaussianProcesses.Periodic — Method.

Periodic(ll::Real, lσ::Real, lp::Real)

Create Periodic with length scale exp(ll), signal standard deviation exp(lσ), and period exp(lp).

source

GaussianProcesses.Poly — Type.

Poly <: Kernel

Polynomial kernel (covariance)

\[k(x,x') = σ²(xᵀx' + c)ᵈ\]

with signal standard deviation $σ$, additive constant $c$, and degree $d$.

source

GaussianProcesses.Poly — Method.

Poly(lc::Real, lσ::Real, deg::Int)

Create Poly with signal standard deviation exp(lσ), additive constant exp(lc), and degree deg.

source

GaussianProcesses.RQ — Method.

RQ(ll::Union{Real,Vector{Real}}, lσ::Real, lα::Real)

Create Rational Quadratic kernel with length scale exp.(ll), signal standard deviation exp(lσ), and shape parameter exp(lα).

See also RQIso and RQArd.

source

GaussianProcesses.RQArd — Type.

RQArd <: StationaryARD{WeightedSqEuclidean}

ARD Rational Quadratic kernel (covariance)

\[k(x,x') = σ²(1 + (x - x')ᵀL⁻²(x - x')/(2α))^{-α}\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$, signal standard deviation $σ$, and shape parameter $α$ where $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.RQArd — Method.

RQArd(ll::Vector{Real}, lσ::Real, lα::Real)

Create RQArd with length scale exp.(ll), signal standard deviation exp(lσ), and shape parameter exp(lα).

source

GaussianProcesses.RQIso — Type.

RQIso <: Isotropic{SqEuclidean}

Isotropic Rational Quadratic kernel (covariance)

\[k(x,x') = σ²(1 + (x - x')ᵀ(x - x')/(2αℓ²))^{-α}\]

with length scale $ℓ$, signal standard deviation $σ$, and shape parameter $α$.

source

GaussianProcesses.RQIso — Method.

RQIso(ll:T, lσ::T, lα::T)

Create RQIso with length scale exp(ll), signal standard deviation exp(lσ), and shape parameter exp(lα).

source

GaussianProcesses.SE — Method.

SE(ll::Union{Real,Vector{Real}}, lσ::Real)

Create squared exponential kernel with length scale exp.(ll) and signal standard deviation exp(lσ).

See also SEIso and SEArd.

source

GaussianProcesses.SEArd — Type.

SEArd <: StationaryARD{WeightedSqEuclidean}

ARD Squared Exponential kernel (covariance)

\[k(x,x') = σ²\exp(- (x - x')ᵀL⁻²(x - y)/2)\]

with length scale $ℓ = (ℓ₁, ℓ₂, …)$ and signal standard deviation $σ$ where $L = diag(ℓ₁, ℓ₂, …)$.

source

GaussianProcesses.SEArd — Method.

SEArd(ll::Vector{Real}, lσ::Real)

Create SEArd with length scale exp.(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.SEIso — Type.

SEIso <: Isotropic{SqEuclidean}

Isotropic Squared Exponential kernel (covariance)

\[k(x,x') = σ²\exp(- (x - x')ᵀ(x - x')/(2ℓ²))\]

with length scale $ℓ$ and signal standard deviation $σ$.

source

GaussianProcesses.SEIso — Method.

SEIso(ll::T, lσ::T)

Create SEIso with length scale exp(ll) and signal standard deviation exp(lσ).

source

GaussianProcesses.EmptyData — Type.

EmptyData <: KernelData

Default empty KernelData.

source

GaussianProcesses.KernelData — Type.

KernelData

Data to be used with a kernel object to calculate a covariance matrix, which is independent of kernel hyperparameters.

See also EmptyData.

source

GaussianProcesses.cov! — Function.

cov!(cK::AbstractMatrix, k::Kernel, X₁::AbstractMatrix, X₂::AbstractMatrix)

Like cov(k, X₁, X₂), but stores the result in cK rather than a new matrix.

source

Statistics.cov — Function.

cov(k::Kernel, X₁::AbstractMatrix, X₂::AbstractMatrix)

Create covariance matrix from kernel k and matrices of observations X₁ and X₂, where each column is an observation.

source

Statistics.cov — Method.

cov(k::Kernel, X::AbstractMatrix[, data::KernelData = KernelData(k, X, X)])

Create covariance function from kernel k, matrix of observations X, where each column is an observation, and kernel data data constructed from input observations.

source